refactor(session): isolate pending tool execution
