+++ opencollab/application/session_run.py
@@ -33,13 +33,8 @@ from opencollab.application.ports import (
     ShaperPort,
     TracePort,
 )
-from opencollab.application.tool_execution import (
-    TERMINAL_CAPTURE_SKIP_MESSAGE,
-    ToolExecutionUseCase,
-)
-from opencollab.domain.pending import PendingEventTable, PendingRow, RowKind, RowStatus
+from opencollab.application.tool_execution import ToolExecutionUseCase
 from opencollab.domain.session import SessionPhase, SessionState
-from opencollab.domain.tools import ToolProcessingResult
 
 __all__ = [
     "DEFAULT_COMMIT_RESERVE",
@@ -670,132 +665,3 @@ class SessionRunUseCase(_SessionRunCompletionMixin):
         await self.finish_step(pending.latency)
         self.clear_pending_step()
         self.state.transition_to(SessionPhase.DONE, reason="completed")
-
-    async def execute_pending_tools(self) -> None:
-        """Run the pending response's tool calls.
-
-        All-synchronous batches run immediately and proceed to AUTOSAVING. A
-        batch containing a deferrable tool (e.g. ``spawn_agent``) is buffered
-        whole into the pending-events table so the eventual tool-result block
-        stays contiguous; the session then suspends on AWAITING_EVENTS until
-        the scheduler delivers the outstanding results.
-        """
-        pending = self._pending
-        if pending is None:
-            self.state.fail()
-            raise RuntimeError("Cannot execute tools before calling LLM")
-
-        original_tool_calls = list(pending.response.tool_calls)
-        preflight = getattr(self.tool_execution, "preflight_tool_batch", None)
-        rejected_batch = getattr(
-            self.tool_execution,
-            "preflight_rejection_result",
-            None,
-        )
-        if callable(preflight) and callable(rejected_batch):
-            preflight_errors = preflight(original_tool_calls)
-            if any(preflight_errors):
-                rejected_batch(
-                    original_tool_calls,
-                    preflight_errors,
-                ).apply_to(self.state)
-                self._pending_tool_allowlist = None
-                self._pending_tool_gate_label = None
-                self.state.transition_to(SessionPhase.AUTOSAVING)
-                return
-        tool_calls, blocked_messages = self._apply_pending_tool_allowlist(original_tool_calls)
-        _immediate, deferred = self._split_tool_calls(tool_calls)
-
-        # Fast path, unchanged: every tool is synchronous — run them, append
-        # results, and autosave the step.
-        if not deferred:
-            result = await self.tool_execution.process(tool_calls) if tool_calls else ToolProcessingResult()
-            result.messages_to_append = self._ordered_tool_messages(
-                original_tool_calls,
-                result.messages_to_append,
-                blocked_messages,
-            )
-            result.apply_to(self.state)
-            self._pending_tool_allowlist = None
-            self._pending_tool_gate_label = None
-            self.state.transition_to(SessionPhase.AUTOSAVING)
-            return
-
-        # A deferred tool is present. Buffer the WHOLE batch into the pending
-        # table so the eventual tool-result block stays contiguous and in the
-        # original order, satisfying the LLM rule that every tool_call_id is
-        # answered before the next model call.
-        table = self.state.pending_events
-        order = {tc["id"]: i for i, tc in enumerate(original_tool_calls)}
-        completed_messages = list(blocked_messages)
-        observations = ToolProcessingResult()
-        terminal_capture_accepted = False
-        for tc in tool_calls:
-            if terminal_capture_accepted:
-                completed_messages.append(
-                    {
-                        "role": "tool",
-                        "tool_call_id": tc["id"],
-                        "content": TERMINAL_CAPTURE_SKIP_MESSAGE,
-                    }
-                )
-                continue
-            if tc.get("function", {}).get("name") in self.deferrable_tool_names:
-                await self._execute_deferred_tools(table, order, [tc])
-                continue
-
-            proc = await self.tool_execution.process([tc])
-            proc.apply_hashes_to(self.state)  # hashes now; messages buffered
-            observations.reads_executed += proc.reads_executed
-            observations.write_succeeded |= proc.write_succeeded
-            observations.read_write_signals.extend(proc.read_write_signals)
-            observations.evidence_signals.extend(proc.evidence_signals)
-            observations.evidence_cards.extend(proc.evidence_cards)
-            observations.loop_detections.extend(proc.loop_detections)
-            observations.tool_step_attempted |= proc.tool_step_attempted
-            completed_messages.extend(proc.messages_to_append)
-            terminal_capture_accepted = proc.terminal_capture_accepted
-
-        # Fold the provider batch once so the progress watchdog remains per-step,
-        # while read/write and evidence signals replay in declared call order.
-        observations.apply_read_write_counter_to(self.state)
-        observations.apply_evidence_counter_to(self.state)
-        self._buffer_completed_rows(table, order, completed_messages)
-
-        self._pending_tool_allowlist = None
-        self._pending_tool_gate_label = None
-        if table.is_complete():
-            # Nothing is actually outstanding (e.g. all spawns were rejected
-            # synchronously) — drain now and autosave instead of suspending on
-            # an event that will never arrive.
-            for message in table.ordered_results():
-                self.state.append_message(message)
-            table.clear()
-            self.state.transition_to(SessionPhase.AUTOSAVING)
-            return
-
-        # Genuinely waiting on ≥1 child. Suspend WITHOUT finishing the step:
-        # emitting step_end here would autosave a half-open tool-call block (no
-        # results yet). The next AUTOSAVING after resume persists a consistent
-        # history. tool_start/tool_end + the scheduler's agent_spawned still
-        # provide observability of the spawn.
-        self.clear_pending_step()
-        self.state.transition_to(SessionPhase.AWAITING_EVENTS)
-
-    @staticmethod
-    def _buffer_completed_rows(
-        table: PendingEventTable,
-        order: dict[str, int],
-        messages: list[dict],
-    ) -> None:
-        for message in messages:
-            tool_call_id = message["tool_call_id"]
-            table.add(
-                PendingRow(
-                    tool_call_id=tool_call_id,
-                    kind=RowKind.IMMEDIATE,
-                    order=order[tool_call_id],
-                    status=RowStatus.DONE,
-                    result=message["content"],
-                )
-            )
